Ai-Integration/patient-features/medical_diagnosis_system/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
imbalanced-learn
joblib
matplotlib
seaborn
shap
xgboost
lightgbm
requests
python-dotenv
